requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
asyncio==4.0.0
asyncio-mqtt==0.16.2
attrs==25.3.0
black==25.9.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.8.0
et_xmlfile==2.0.0
fonttools==4.60.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.182.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.75.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
isort==6.0.1
Jinja2==3.1.6
jiter==0.11.0
json-logging==1.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.6
mdurl==0.1.2
motor==3.7.1
mypy==1.18.2
mypy_extensions==1.1.0
narwhals==2.5.0
numpy==2.3.3
openai==1.108.1
openpyxl==3.1.5
packaging==25.0
paho-mqtt==2.1.0
pandas==2.3.2
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
plotly==6.3.0
pluggy==1.6.0
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pymongo==4.15.1
pyparsing==3.2.5
pytest==8.4.2
pytest-asyncio==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytokens==0.1.10
pytz==2025.2
referencing==0.36.2
requests==2.32.5
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
scipy==1.16.2
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.49.1
structlog==25.4.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uuid==1.30
watchdog==6.0.0
